print sum of even value Fibonacci numbers
sum up even values of Fibonacci numbers up to 4000000

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int i = 1, j = 2, k = 0, sum = 0;
+	long int i = 1, j = 2, k = 0, sum = 0;
 	while (j <= 4000000)
 	{
 		if ((j % 2) == 0)
